+++ src/ui_helpers.py
@@ -1,0 +1,27 @@
+import customtkinter
+import tkinter
+# import logging
+
+# -----------------------------------------------------------------------------
+
+def textbox_find_text(textbox: customtkinter.CTkTextbox, needle: str) -> int:
+    """Search and highlight the text in the TextBox"""
+
+    textbox.tag_delete("search")
+    textbox.tag_config("search", background="yellow")
+    start = 1.0
+    pos = textbox.search(pattern=needle, index=start, stopindex=tkinter.END, nocase=True)
+    found_cnt = 0
+    length = len(needle)
+
+    if length > 0:
+        while pos:
+            found_cnt += 1
+            row, col = pos.split('.')
+            end = int(col) + length
+            end = f'{row}.{str(end)}'
+            textbox.tag_add("search", pos, end)
+            start = end
+            pos = textbox.search(pattern=needle, index=start, stopindex=tkinter.END, nocase=True)
+
+    return found_cnt