+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 ## 0.1.0 - 2023-07-31
 
 * Added
+  * application window build with customtkinter
+  * PnP parser and preview from the Boomer project
 * Changed
 * Deprecated
 * Removed